fix(tower): meshmessage/feedbackentry integer-decode panic (empty reply)
Found live in production while verifying the #111 progress-seq fix: the
GET /work/tasks/{id}/progress endpoint returned "Empty reply from server"
on every call. Tower logs showed a real thread panic at sqlx-core's
row.rs:72:37 (Row::get's internal try_get().unwrap()), crashing the
connection.

Root cause, and how far it reaches: this schema uses `integer` (i32) for
every `id` column, but several handlers decoded them as i64 (or as
non-Option when the underlying column is nullable) via Row::get, which
panics on any type mismatch instead of erroring gracefully. Confirmed via
the schema (crates/edgeplane-tower/migrations/0001_initial_schema.sql) and
fixed every site this class of bug actually reaches:

- routes/work.rs `get_task_progress` (the endpoint that surfaced this):
  meshprogressevent.id decoded i64->i32; summary/payload_json (nullable
  text) decoded as non-Option, now default gracefully.
- routes/work.rs `row_to_message` (used by list_domain_messages,
  list_mission_messages, get_agent_messages -- the actual
  send_mesh_message/list_mesh_messages read path this whole fleet's
  inter-agent messaging runs on): meshmessage.id and .in_reply_to
  (nullable integer) decoded i64->i32; body_json (nullable text) decoded
  as non-Option, now defaults gracefully.
- routes/work.rs `send_domain_message`/`send_mission_message`: the
  INSERT...RETURNING id decode, same i64->i32 fix.
- routes/work.rs `get_agent_messages`' read-marking direct_ids collection:
  Vec<i64> -> Vec<i32> to match.
- routes/feedback.rs `row_to_feedback`: feedbackentry.id, same fix.

This predates today's work entirely -- unrelated to #108/#110/#111's
changes, just found by chance while live-verifying #111 against the real
tower. Messaging (send/list) was live-crashing on any message with a
populated in_reply_to or non-trivial body_json; feedback listing was
crashing on every row.

New regression test drives the actual send+list HTTP path (not just the
decode in isolation) and asserts on the in_reply_to linkage and body_json
round-trip, since those are exactly the fields that were panicking.

Verified: tower suite 200/200 (was crashing this exact request path
moments before the fix, confirmed via live curl + tower pod logs before
touching code); clippy --workspace --all-targets -D warnings clean; CLI
199/199; workspace 499/499; launch-regression ok; no doc/dockerfile
drift.

+++ crates/edgeplane-tower/src/routes/feedback.rs
@@ -78,7 +78,9 @@ fn row_to_feedback(row: &sqlx::postgres::PgRow) -> serde_json::Value {
         serde_json::from_str(row.get::<&str, _>("metadata_json"))
             .unwrap_or(serde_json::json!({}));
     serde_json::json!({
-        "id": row.get::<i64, _>("id"),
+        // feedbackentry.id is `integer` (i32) in Postgres — decoding as i64 panics
+        // via `Row::get`, crashing the request with an empty reply.
+        "id": row.get::<i32, _>("id"),
         "domain_id": row.get::<String, _>("domain_id"),
         "mission_id": row.get::<String, _>("mission_id"),
         "source_type": row.get::<String, _>("source_type"),
